requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aiosignal==1.3.2
airportsdata==20250523
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blake3==1.0.5
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
compressed-tensors==0.9.4
cupy-cuda12x==13.4.1
datasets==3.6.0
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
flashinfer-python==0.2.5
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.0
gguf==0.17.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
interegular==0.3.3
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
lark==1.2.2
llguidance==0.7.26
llmcompressor==0.5.1
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistral_common==1.5.6
mmh3==5.1.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
narwhals==1.41.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
ollama==0.5.1
onnxruntime==1.22.0
openai==1.82.1
opencv-python-headless==4.11.0.86
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-semantic-conventions-ai==0.4.9
opentelemetry-util-http==0.48b0
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post5
pillow==11.2.1
posthog==4.2.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyMuPDF==1.26.0
pymupdf4llm==0.0.24
pynvml==12.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
# Editable install with no version control (RAG_Agent_v1==0.0.0)
-e /home/mark/projects/Canada-Labour-Research-Assistant/vllm_setup
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.45.3
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uv==0.7.9
uvicorn==0.34.3
uvloop==0.21.0
vllm==0.9.0.1
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
